Added water 'suction'
added suction effect to invisible death object. It only suctions the
player, no enemies are affected. Configuration variables are avaliable
in the inivisible death object create event. The suction code is in the
player collision event with invisible death object. Note: it uses the
knockback system to pull, that could be changed and it hasn't been
tested fully with knockback. It may behave unpredicatably if the death
object is too small to allow for the pull border.